requirements.txt
fastapi
jinja2
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
python-multipart
slowapi
itsdangerous